Fix clang warning. The RNG function used in wagon is biased.
1. The conversion from RAND_MAX to float implies a bias
   because RAND_MAX (2^31-1) can't be represented exactly
   as a float.

2. C++11 provides a cleaner alternative. Add a comment
   with respect to that, in case speech-tools ever migrates
   to C++11 or newer.

+++ stats/wagon/wagon.cc
@@ -810,7 +810,12 @@ void wgn_find_split(WQuestion &q,WVectorVector &ds,
 static float wgn_random_number(float x)
 {
     // Returns random number between 0 and x
-    return (((float)random())/RAND_MAX)*x;
+    /* This is not a good approach for generating random
+     * numbers. The conversion of RAND_MAX from int to float
+     * implies a loss of precision. C++-11 has std::uniform_real_distribution
+     * that provides a cleaner solution. Consider adopting c++-11
+     */
+    return (((float)random())/(float)RAND_MAX)*x;
 }
 
 #ifdef OMP_WAGON
